Add Reset Dialogs:
A reset dialog, which will pop up by pressing the reset menu button,
is already used for Quick settings and nav bar settings,
to be consistent, I add a reset dialog to all additional settings
which already have a reset menu.
This will also change the 3 dot icon to the backup icon which is used
for the Quick settings and nav bar settings reset menu.

Also clean up the code.